+++ packages/ragbits-core/src/ragbits/core/embeddings/vertex_multimodal.py
@@ -1,6 +1,8 @@
 import asyncio
 import base64
 
+from ragbits.core.embeddings.litellm import LiteLLMEmbeddingsOptions
+
 try:
     import litellm
     from litellm.llms.vertex_ai_and_google_ai_studio.common_utils import VertexAIError
@@ -18,11 +20,12 @@ from ragbits.core.embeddings.exceptions import (
 )
 
 
-class VertexAIMultimodelEmbeddings(Embeddings):
+class VertexAIMultimodelEmbeddings(Embeddings[LiteLLMEmbeddingsOptions]):
     """
     Client for creating text embeddings using LiteLLM API.
     """
 
+    options_cls = LiteLLMEmbeddingsOptions
     VERTEX_AI_PREFIX = "vertex_ai/"
 
     def __init__(
@@ -31,7 +34,7 @@ class VertexAIMultimodelEmbeddings(Embeddings):
         api_base: str | None = None,
         api_key: str | None = None,
         concurency: int = 10,
-        options: dict | None = None,
+        default_options: LiteLLMEmbeddingsOptions | None = None,
     ) -> None:
         """
         Constructs the embedding client for multimodal VertexAI models.
@@ -41,7 +44,7 @@ class VertexAIMultimodelEmbeddings(Embeddings):
             api_base: The API endpoint you want to call the model with.
             api_key: API key to be used. If not specified, an environment variable will be used.
             concurency: The number of concurrent requests to make to the API.
-            options: Additional options to pass to the API.
+            default_options: Additional options to pass to the API.
 
         Raises:
             ImportError: If the 'litellm' extra requirements are not installed.
@@ -50,7 +53,7 @@ class VertexAIMultimodelEmbeddings(Embeddings):
         if not HAS_LITELLM:
             raise ImportError("You need to install the 'litellm' extra requirements to use LiteLLM embeddings models")
 
-        super().__init__()
+        super().__init__(default_options=default_options)
         if model.startswith(self.VERTEX_AI_PREFIX):
             model = model[len(self.VERTEX_AI_PREFIX) :]
 
@@ -58,19 +61,19 @@ class VertexAIMultimodelEmbeddings(Embeddings):
         self.api_base = api_base
         self.api_key = api_key
         self.concurency = concurency
-        self.options = options or {}
 
         supported_models = VertexMultimodalEmbedding().SUPPORTED_MULTIMODAL_EMBEDDING_MODELS
         if model not in supported_models:
             raise ValueError(f"Model {model} is not supported by VertexAI multimodal embeddings")
 
-    async def _embed(self, data: list[dict]) -> list[dict]:
+    async def _embed(self, data: list[dict], options: LiteLLMEmbeddingsOptions | None = None) -> list[dict]:
         """
         Creates embeddings for the given data. The format is defined in the VertexAI API:
         https://cloud.google.com/vertex-ai/generative-ai/docs/embeddings/get-multimodal-embeddings
 
         Args:
             data: List of instances in the format expected by the VertexAI API.
+            options: Additional options to pass to the VertexAI multimodal embeddings API.
 
         Returns:
             List of embeddings for the given VertexAI instances, each instance is a dictionary
@@ -80,16 +83,17 @@ class VertexAIMultimodelEmbeddings(Embeddings):
             EmbeddingStatusError: If the embedding API returns an error status code.
             EmbeddingResponseError: If the embedding API response is invalid.
         """
+        merged_options = (self.default_options | options) if options else self.default_options
         with trace(
             data=data,
             model=self.model,
             api_base=self.api_base,
-            options=self.options,
+            options=merged_options.dict(),
         ) as outputs:
             semaphore = asyncio.Semaphore(self.concurency)
             try:
                 response = await asyncio.gather(
-                    *[self._call_litellm(instance, semaphore) for instance in data],
+                    *[self._call_litellm(instance, semaphore, merged_options) for instance in data],
                 )
             except VertexAIError as exc:
                 raise EmbeddingStatusError(exc.message, exc.status_code) from exc
@@ -102,13 +106,16 @@ class VertexAIMultimodelEmbeddings(Embeddings):
 
             return outputs.embeddings
 
-    async def _call_litellm(self, instance: dict, semaphore: asyncio.Semaphore) -> litellm.EmbeddingResponse:
+    async def _call_litellm(
+        self, instance: dict, semaphore: asyncio.Semaphore, options: LiteLLMEmbeddingsOptions
+    ) -> litellm.EmbeddingResponse:
         """
         Calls the LiteLLM API to get embeddings for the given data.
 
         Args:
             instance: Single VertexAI instance to get embeddings for.
             semaphore: Semaphore to limit the number of concurrent requests.
+            options: Additional options to pass to the VertexAI multimodal embeddings API.
 
         Returns:
             List of embeddings for the given LiteLLM instances.
@@ -119,17 +126,18 @@ class VertexAIMultimodelEmbeddings(Embeddings):
                 model=f"{self.VERTEX_AI_PREFIX}{self.model}",
                 api_base=self.api_base,
                 api_key=self.api_key,
-                **self.options,
+                **options.dict(),
             )
 
         return response
 
-    async def embed_text(self, data: list[str]) -> list[list[float]]:
+    async def embed_text(self, data: list[str], options: LiteLLMEmbeddingsOptions | None = None) -> list[list[float]]:
         """
         Creates embeddings for the given strings.
 
         Args:
             data: List of strings to get embeddings for.
+            options: Additional options to pass to the VertexAI multimodal embeddings API.
 
         Returns:
             List of embeddings for the given strings.
@@ -138,7 +146,7 @@ class VertexAIMultimodelEmbeddings(Embeddings):
             EmbeddingStatusError: If the embedding API returns an error status code.
             EmbeddingResponseError: If the embedding API response is invalid.
         """
-        response = await self._embed([{"text": text} for text in data])
+        response = await self._embed([{"text": text} for text in data], options=options)
         return [embedding["textEmbedding"] for embedding in response]
 
     def image_support(self) -> bool:  # noqa: PLR6301
@@ -150,12 +158,15 @@ class VertexAIMultimodelEmbeddings(Embeddings):
         """
         return True
 
-    async def embed_image(self, images: list[bytes]) -> list[list[float]]:
+    async def embed_image(
+        self, images: list[bytes], options: LiteLLMEmbeddingsOptions | None = None
+    ) -> list[list[float]]:
         """
         Creates embeddings for the given images.
 
         Args:
             images: List of images to get embeddings for.
+            options: Additional options to pass to the VertexAI multimodal embeddings API.
 
         Returns:
             List of embeddings for the given images.
@@ -165,6 +176,8 @@ class VertexAIMultimodelEmbeddings(Embeddings):
             EmbeddingResponseError: If the embedding API response is invalid.
         """
         images_b64 = (base64.b64encode(image).decode() for image in images)
-        response = await self._embed([{"image": {"bytesBase64Encoded": image}} for image in images_b64])
+        response = await self._embed(
+            [{"image": {"bytesBase64Encoded": image}} for image in images_b64], options=options
+        )
 
         return [embedding["imageEmbedding"] for embedding in response]